cleans up tests
